Use OpenCL data types

+++ matmul/matmul.cl
@@ -1,15 +1,15 @@
 __kernel
 void matmul(__constant double *Mvals,
-            __constant unsigned int *Mrow_ptr,
-            __constant unsigned int *Mcol_idx,
+            __constant uint   *Mrow_ptr,
+            __constant uint   *Mcol_idx,
             __constant double *V,
             __global   double *Ret)
 {
-   const unsigned int tid = get_global_id(0);
+   const size_t tid = get_global_id(0);
 
    Ret[tid] = 0.0;
 
-   for (unsigned int j=Mrow_ptr[tid]; j<Mrow_ptr[tid+1]; ++j) {
+   for (uint j=Mrow_ptr[tid]; j<Mrow_ptr[tid+1]; ++j) {
       Ret[tid] += Mvals[j] * V[Mcol_idx[j]];
    }
 }
